pyproject.toml
[tool.poetry]
name = "advent-of-code-2023"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.0"
advent-of-code-data = "^2.0.1"
numpy = "^1.26.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
